pyproject.toml
[project]
name = "teaching_assistant_crew"
version = "0.1.0"
description = "AI-powered Teaching Assistant using CrewAI and Ollama"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "crewai[tools]>=1.9.0,<2.0.0",
    "litellm>=1.55,<1.60",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.100.0",
    "twilio>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
teaching-assistant = "ollama_swarm.main:run"
ta-daily = "ollama_swarm.main:run_daily_briefing"
ta-ingest = "ollama_swarm.main:run_email_ingest"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ollama_swarm"]
